crashes in debug

+++ lib/code_tree.cpp
@@ -9,7 +9,7 @@ code_tree::code_tree() {
     decode_num = &code_tree::decode_by_tree;
 }
 
-code_tree::code_tree(std::unique_ptr<size_t[]> &freq) {
+code_tree::code_tree(std::unique_ptr<uint64_t[]> &freq) {
     decode_num = &code_tree::decode_by_tree;
     struct node_cmp {
         bool operator()(node *lhs, node *rhs) const {
@@ -19,7 +19,7 @@ code_tree::code_tree(std::unique_ptr<size_t[]> &freq) {
 
     auto queue = std::priority_queue<node *, std::vector<node *>, node_cmp>();
 
-    for (size_t i = 0; i < SYMBOL_CNT; ++i) {
+    for (uint64_t i = 0; i < SYMBOL_CNT; ++i) {
         if (freq[i] != 0) {
             queue.push(new node(static_cast<uint16_t>(i), freq[i]));
         }
@@ -54,13 +54,13 @@ symbol code_tree::decode_by_tree(uint64_t code_piece) {
 }
 
 code_tree::node *
-code_tree::make_node(std::vector<std::pair<int32_t, int32_t>> const &init_data, size_t i) { //!!! MB INFINITE RECURSION
+code_tree::make_node(std::vector<std::pair<int32_t, int32_t>> const &init_data, uint64_t i) { //!!! MB INFINITE RECURSION
     node *res, *ch1, *ch2;
     if (init_data[i].second == -1) {
         res = new node(static_cast<uint16_t>(init_data[i].first));
     } else {
-        ch1 = make_node(init_data, static_cast<size_t>(init_data[i].first));
-        ch2 = make_node(init_data, static_cast<size_t>(init_data[i].second));
+        ch1 = make_node(init_data, static_cast<uint64_t>(init_data[i].first));
+        ch2 = make_node(init_data, static_cast<uint64_t>(init_data[i].second));
         res = new node(ch2, ch1);
     }
     return res;
@@ -134,7 +134,7 @@ void code_tree::switch_to_table_mode() {
         uint8_t max_code_length = 0;
         if (cheat_table == nullptr) {
             cheat_table = new std::pair<symbol, uint8_t>[CHEAT_TABLE_LENGTH];
-            for (size_t i = 0; i < CHEAT_TABLE_LENGTH; ++i) {
+            for (uint64_t i = 0; i < CHEAT_TABLE_LENGTH; ++i) {
                 cheat_table[i].first = decode_by_tree(i);
                 cheat_table[i].second = code_pos;
                 max_code_length = std::max(max_code_length, code_pos);
